+++ src/ber/parser.rs
@@ -1172,12 +1172,6 @@ pub fn parse_ber_recursive(i: &[u8], max_depth: usize) -> BerResult {
     }
     match hdr.class() {
         Class::Universal => (),
-        Class::Private => {
-            let (rem, content) = ber_get_object_content(rem, &hdr, max_depth)?;
-            let content = BerObjectContent::Private(hdr.clone(), content);
-            let obj = BerObject::from_header_and_content(hdr, content);
-            return Ok((rem, obj));
-        }
         _ => {
             let (rem, data) = ber_get_object_content(rem, &hdr, max_depth)?;
             let any = Any::new(hdr.clone(), data);